safari-600.7-branch/LayoutTests/http/tests/appcache/resource-redirect-2-expected.txt

CONSOLE MESSAGE: Application Cache update failed, because http://127.0.0.1:8000/appcache/resources/resource-redirect-2.php could not be fetched.
Test that a redirect makes resource caching fail.

Should say SUCCESS:

SUCCESS
